refactor utils in model package into separate concerns.
